requirements.txt
 pycryptodome==3.18.0 
 cryptography==41.0.1
 xxtea==3.0.0